backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai-whisper
torch
dashscope
moviepy
python-multipart
pyjwt
passlib[bcrypt]
